feat: extract book card list component
