+++ src/pages/HomePage/components/SearchPanel/SearchPanel.tsx
@@ -1,5 +1,6 @@
 import { Component, ChangeEvent } from 'react';
-import { BookCardResponseBooks } from '../models';
+import { BookCardResponseBooks } from '../BookCard/models';
+import './SearchPanel.css';
 
 interface State {
   searchTerm: string;
@@ -20,6 +21,7 @@ class SearchPanel extends Component<object, State> {
       .then((response) => response.json())
       .then((data) => {
         console.log(data);
+        console.log(this.state);
         this.setState({ searchResults: [data] });
       })
       .catch((error) => {
@@ -27,14 +29,14 @@ class SearchPanel extends Component<object, State> {
       });
   };
 
-  handleSearchInputChange = (e: ChangeEvent<HTMLInputElement>) => {
-    this.setState({ searchTerm: e.target.value });
+  handleSearchInputChange = (event: ChangeEvent<HTMLInputElement>) => {
+    this.setState({ searchTerm: event.target.value });
   };
 
   render() {
     return (
       <div>
-        <div className="top-section">
+        <div className="search-box">
           <input
             type="text"
             value={this.state.searchTerm}
@@ -43,7 +45,7 @@ class SearchPanel extends Component<object, State> {
           />
           <button onClick={this.handleSearch}>Search</button>
         </div>
-        <div className="bottom-section">
+        <div className="card-list">
           {this.state.searchResults.map((result, index) => (
             <div key={index}>
               <h2>{result.title}</h2>
